+++ solver/Cargo.toml
@@ -9,3 +9,5 @@ edition = "2021"
 dfpn = { path = "../dfpn" }
 dfpn-extended = { path = "../dfpn-extended" }
 shogi_core = "0.1.4"
+shogi_usi_parser = "0.1.0"
+yasai = { git = "https://github.com/sugyan/yasai", tag = "0.5.0", features = ["simd"] }
