+++ solver/src/implementations.rs
