+++ solver/src/implementations/hashmap_table.rs
