+++ solver/src/implementations/vec_table.rs
