+++ solver/src/implementations/yasai_position.rs
@@ -1,6 +1,6 @@
+use crate::CalculateResult;
 use dfpn::Node;
 use shogi_core::{Hand, Move, PartialPosition};
-use solver::CalculateResult;
 use yasai::Position;
 
 pub struct YasaiPosition(Position);
