+++ src/main.rs
@@ -5,12 +5,12 @@ use shogi_converter::Record;
 use shogi_core::{Color, Hand, Move, PartialPosition, PieceKind, Square, ToUsi};
 use shogi_official_kifu::display_single_move_kansuji;
 use shogi_usi_parser::FromUsi;
+use solver::implementations::{HashMapTable, YasaiPosition};
 use solver::solve;
 use std::fs::File;
 use std::io::{BufRead, Read};
 use std::time::{Duration, Instant};
 use thiserror::Error;
-use tsumeshogi_solver::implementations::{HashMapTable, YasaiPosition};
 
 #[derive(Error, Debug)]
 enum ParseError {
